Handle list of directories in $GOPATH
When $GOPATH is a list of directories, e.g. colon separated on *nix,
both path.GoSrcDir and path.GomacroDir are set incorrectly.

GoSrcDir is used to place temporary files as a result of import
statements. If $GOPATH is a list of directories you end up with
erroneous directories like /a:/b:/c/src/gomacro_imports/{pkg} in the
filesystem.

GomacroDir is currently set to GoSrcDir/src/github.com/cosmos72/gomacro.
When $GOPATH is a list, that directory never exists. GomacroDir is used
to place temporary files (for importing purposes) in the same
directory of gomacro so that it can be recompiled to use new imports.

This diff sets GoSrcDir to the first directory in $GOPATH, and sets
GomacroDir to the correct directory by scanning $GOPATH. If gomacro
source code is not available, prints a warning and sets it to GoSrcDir.

I'm not strong about the GomacroDir behavior and previously had a panic
when gomacro source is not available. However, enforcing users to carry
gomacro source code is probably not desirable, especially for systems
that can handle imports via Go plugins, like Linux and Mac.

This diff fix all import modes except ImInception. The ImInception mode
blindly assumes the imported package's source code is in GoSrcDir as
opposed to scanning $GOPATH.

Closes #32