Support flattening AstWithSlice on Append
Prior to the change, the Append method of NodeSlice would panic if the
child satified AstWithSlice itself.  The panic originates from the
ToNode call, since none of the AstWithSlice types also satify
AstWithNode.

This change simply makes a call to ToNodesAppend instead of append.
ToNodesAppend already checks for AstWithSlice and will unwrap the
children.  Thus this allows one to append a slice into another slice,
resulting in one flat slice.