Introduce CLI version option
Prior to this change, gomacro was unable to report its own version.

This change introduces the cmd.Version package variable and Cmd.Version
method, together these print the binary's version, which can be injected
at build time.  The `-V` and `--version` options are now recognised by
the cmd.Main method and these options, as well as `-h` and `--help`,
short-circuit all other options on the command line.  This avoids
evaluating any possible expressions and the like if the user just wants
a quick version or usage message.

When releasing the binary, it should be built with something like:

    go build -ldflags="-X github.com/cosmos72/gomacro/cmd.Version=$(git describe --dirty)" main.go

Signed-off-by: Corin Lawson <corin@responsight.com>