Ensure genimport has a name under ImInception mode
Prior to this change, the previous commit to base/genimport/genimport.go
(30b1253a) remove assigning a name to the genimport struct in the
newGenImport function.  This name is essential when the ImInception mode
is used, without it the generated x_package.go has an incomplete package
statement (i.e. the package name is missing).  This leads to invalid
code being generated.

This change reintroduces the assignment to genimport's name within the
newGenImport function.  The assignment only occurs under the ImInception
mode.  I assume that the collectPackageImportsWithRename function is
sufficient for the other modes.