Collect ast.Spec productions
Prior to this change the Globals#CollectNode failed to handle values
that satisfied the ast.Spec interface.  This resulted in missing type
stanzas in the generated code from certain macros.  Note that the issue
could be worked around by returning a []ast.Node (who knows why?)
The issue was quickly identified after adding a default switch clause
that then panicked with the node's type.

This change adds a ast.Spec case clause to the top level switch
statement in CollectNode.  This case clause wraps the spec in a
ast.GenDecl value and recursively calls CollectNode to handle the node.
The GenDecl's TokPos is simply set to the node's Pos in the absence of
more complete information.  ValueSpec is assumed to be a VarSpec (not a
ConstSpec).

Also, panics are added when unrecognised values are offered to the
various collect methods, in the hope that similar issues are caught more
quickly in future.

Fixes #84